vercel_python/requirements.txt
fastapi==0.104.1
openai==1.3.5
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
sqlalchemy==2.0.23
pydantic==2.5.2
linkedin-api==2.3.1